Allow BatchResultsImpl to use a supertype ID function.
Previously the Function had to convert from the exact same type as the object type, but now with <? super T>
the function can declare a supertype of T, e.g. for a HtmlPage resultset you can use a function that works
on a Page, since all HtmlPages will be Pages.