Fix some broken tests.
I'd upgraded JUnit and JMock to try mocking out some classes,
but then I no longer needed to do that mocking. But looks like
the newer JUnit and/or JMock cause some of the tests to fail
in ways that are beyond mortal comprehension, so I've done the
sensible thing by reverting back to the previous versions.